add the function of not doing anything to temp_dl

+++ botutils/packages/files.py
@@ -10,12 +10,15 @@ class TempDownload:
     def __init__(self, filename: str, url: str):
         self.filename = filename
         self.url = url
-        files = os.listdir("./.local")
-        if filename in files:
-            self.filename += str(len([f for f in files if filename in f]))
-        self.fp = os.path.join(os.getcwd(), ".local", self.filename)
+        files = os.listdir("./.temp")
+        if filename:
+            if filename in files:
+                self.filename += str(len([f for f in files if filename in f]))
+            self.fp = os.path.join(os.getcwd(), ".temp", self.filename)
 
     async def __aenter__(self):
+        if not self.filename:
+            return None
         async with aiohttp.ClientSession() as sess:
             async with sess.get(self.url) as resp:
                 raw_dat = await resp.read()
@@ -24,6 +27,8 @@ class TempDownload:
         return self.fp
 
     async def __aexit__(self, _exc_type, _exc_val, _exc_tb):
+        if not self.filename:
+            return None
         if os.path.isfile(self.fp):
             os.remove(self.fp)
